Updated delete data menu to display dialog to ask for confirmation.

+++ app/src/main/java/com/digitalruiz/pizzadriver/settings.java
@@ -1,9 +1,13 @@
 package com.digitalruiz.pizzadriver;
 
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.content.DialogInterface;
+import android.content.Intent;
 import android.os.Bundle;
 import android.view.MenuItem;
+import android.widget.Toast;
 
 import com.google.android.material.navigation.NavigationView;
 
@@ -25,22 +29,31 @@ public class settings extends AppCompatActivity {
         MenuItem clear_data = nav_draw_settings.getMenu().findItem(R.id.clear_data_menu).setVisible(true);
 
         clear_data.setOnMenuItemClickListener(item -> {
-            pizzaDriverDB.deleteData();
+            AlertDialog.Builder builder = new AlertDialog.Builder(this);
+            builder.setTitle("Delete Data");
+            builder.setMessage("Are you sure you want to delete all data?");
+            builder.setCancelable(false);
+            builder.setPositiveButton("Yes", (dialog, which) -> {
+                pizzaDriverDB.deleteData();
+                Toast.makeText(settings.this, "All data has been deleted", Toast.LENGTH_LONG).show();
+                finish();
+                Intent intent = new Intent(settings.this, MainActivity.class);
+                startActivity(intent);
+            });
+            builder.setNegativeButton("No", (dialog, which) -> {
+                dialog.cancel();
+                Toast.makeText(settings.this, "Ok, not deleting", Toast.LENGTH_SHORT).show();
+                finish();
+
+            });
+            AlertDialog alert = builder.create();
+            alert.show();
             return false;
         });
 
 
 
-
-        //clear_button = findViewById(R.id.clear_data_menu);
-        //clear_data.setOnClickListener(new View.OnClickListener() {
-        //    @Override
-        //    public void onClick(View v) {
-        //        pizzaDriverDB.deleteData();
-        //    }
-        //});
     }
-
 
 
 }
